+++ SnowmanTests/TranslationTests.swift
@@ -35,6 +35,15 @@ class TranslationTests: XCTestCase {
 
         XCTAssertEqual("Greetings".localizedExpectation,
                        "Greetings".localizedValue)
+
+        XCTAssertEqual("Label: Made with Love".localizedExpectation,
+                       "Label: Made with Love".localizedValue)
+
+        XCTAssertEqual("Label: MeteoDataFrom".localizedExpectation,
+                       "Label: MeteoDataFrom".localizedValue)
+
+        XCTAssertEqual("Label: GeoCodingService".localizedExpectation,
+                       "Label: GeoCodingService".localizedValue)
     }
 
     // MARK: - Popover Screen
@@ -48,12 +57,6 @@ class TranslationTests: XCTestCase {
 
         XCTAssertEqual("Button: Call Forecast".localizedExpectation,
                        "Button: Call Forecast".localizedValue)
-
-        XCTAssertEqual("Label: Made with Love".localizedExpectation,
-                       "Label: Made with Love".localizedValue)
-
-        XCTAssertEqual("Label: MeteoDataFrom".localizedExpectation,
-                       "Label: MeteoDataFrom".localizedValue)
 
         XCTAssertEqual("Tab: Current Weather".localizedExpectation,
                        "Tab: Current Weather".localizedValue)
@@ -331,6 +334,7 @@ class TranslationTests: XCTestCase {
 
     // MARK: - Options Screen
 
+    // swiftlint:disable:next function_body_length
     func test_translation_of_OptionsScreen() {
 
         // assert
@@ -367,6 +371,15 @@ class TranslationTests: XCTestCase {
 
         XCTAssertEqual("Option: Distance".localizedExpectation,
                        "Option: Distance".localizedValue)
+
+        XCTAssertEqual("Option: Current Provider".localizedExpectation,
+                       "Option: Current Provider".localizedValue)
+
+        XCTAssertEqual("Option: For Meteo Data".localizedExpectation,
+                       "Option: For Meteo Data".localizedValue)
+
+        XCTAssertEqual("Option: For Geo Data".localizedExpectation,
+                       "Option: For Geo Data".localizedValue)
 
         XCTAssertEqual("Unit: Light".localizedExpectation,
                        "Unit: Light".localizedValue)
@@ -605,6 +618,12 @@ class TranslationTests: XCTestCase {
                        "Label: RU Expectation".localizedValue)
         XCTAssertEqual("Label: Author".localizedExpectation,
                        "Label: Author".localizedValue)
+        XCTAssertEqual("Label: Meteo API".localizedExpectation,
+                       "Label: Meteo API".localizedValue)
+        XCTAssertEqual("Label: Geo API".localizedExpectation,
+                       "Label: Geo API".localizedValue)
+        XCTAssertEqual("Label: Major Components".localizedExpectation,
+                       "Label: Major Components".localizedValue)
         XCTAssertEqual("Button: License".localizedExpectation,
                        "Button: License".localizedValue)
         XCTAssertEqual("Button: Terms & Conditions".localizedExpectation,